changed from f64 to f32 on ln 12 and 17

+++ src/main.rs
@@ -9,11 +9,11 @@ fn main() {
         .read_line(&mut temp_f)
         .expect("Please enter a number.");
 
-    let temp_f: f64 = match temp_f.trim().parse() {
+    let temp_f: f32 = match temp_f.trim().parse() {
         Ok(num) => num,
         Err(_) => 0.0,
     };
 
-    let temp_c: f64 = ({temp_f} - 32.0) * 0.556;
+    let temp_c: f32 = ({temp_f} - 32.0) * 0.556;
     println!("The temperature in centigrade is: {temp_c}");
 }
